createOrg is now working
Made it so users can create an org/ leaderboard. And they show up on the database.

Fixed the directories surrounding createOrg

Made it so users can include blank spaces in org names.

fixed the css a bit.

also made it so users get feedback if their orgs get created or not